+++ packages/swirl-components/src/components/flip-avatar-group/flip-avatar-group.spec.tsx
@@ -20,7 +20,7 @@ describe("flip-avatar-group", () => {
           <div class="avatar-group avatar-group--has-badge" role="group">
             <slot></slot>
             <span class="avatar-group__badge">
-              <flip-badge aria-label="3 new messages" label="3"></flip-badge>
+              <flip-badge aria-label="3 new messages" label="3" size="m"></flip-badge>
             </span>
           </div>
         </mock:shadow-root>